usr.sbin/bhyve: incremental improvement for IO queue handling
The previous implementation assumes only one queue and is performing
pointer arithmetic on a singleton.  Push the handling of arrays up to
the caller.  When these devices are updated to support multiple virtual
queues and virtio_const's vc_nvq is changed from 1, only the callers of
vi_softc_linkup() will need to be revisited.

Coverity CID: 1193568